Make it possible to link NativeAOT outputs
Fixes issues with missing symbols.

Take out a piece of the type loader that requires System.Private.TypeLoader.Native that we didn't bring over for now.

isa_detection.cpp doesn't build in NativeAOT because it doesn't include any headers that would define things like WszLoadLibraryEx or LOAD_LIBRARY_SEARCH_SYSTEM32. It feels like this needs to be factored to use GCToOSInterface but this is not my area of expertise.

Remove reference to __fail_fast that was deleted as part of getting rid of cppcodegen in #21.