Thunks: support for fixed size thunk pool (#8304)
* Thunks: support for fixed size thunk pool

Provides the platform independent (OS and CPU architecture) runtime code to support a fixed size thunk pool for systems that doesn't support any way to add more thunks on demand.

Activated with FEATURE_FIXED_POOL_THUNKS..

FEATURE_RX_THUNKS will still override this.

The default is still the template based system.

The code blocks of the pool itself need to be either build and linked as part of the executable or generated by the compiler to be included in the object file.

* fixed size thunk pool (#23)

* Thunks: support for fixed size thunk pool

Provides the platform independent (OS and CPU architecture) runtime code to support a fixed size thunk pool for systems that doesn't support any way to add more thunks on demand.

Activated with FEATURE_FIXED_POOL_THUNKS..

FEATURE_RX_THUNKS will still override this.

The default is still the template based system.

The code blocks of the pool itself need to be either build and linked as part of the executable or generated by the compiler to be included in the object file.

* Review changes

- Use PalVirtualAlloc instead of GCToOSInterface for memory operations
- Rename global variable to better match the usage