ARM64: fix loading function address in CallingConventionConverter_GetStubs (#8310)
The current way to load the address of a functions there can cause the assembler to add a data field in a executable section itself. Depending on how smart the linker is this might generate position dependent code. In this case the OS might need to patch these data fields after loading. Therefore the linker needs to add relocation information for the code section. A hardened OS can reject to load an file that contains relocation in the executable part for security reasons. 
The new way doesn't have this issue. It is already used for other functions.