ARM64: FEATURE_USE_SOFTWARE_WRITE_WATCH_FOR_GC_HEAP (#8331)
Implements the missing parts of the software write watch system. Activates and deactivates the tracking based on the GC state. Code is based on the same function in the runime

Generates an compiler error if FEATURE_USE_SOFTWARE_WRITE_WATCH_FOR_GC_HEAP is enabled for anything else than ARM64 UNIX as all other assembler helpers are not updated to support it.