Use new GVM generation logic in the scanner (#107)
Move the logic to find slot defining method of generic virtual methods to a common location and use it from the scanner.

The fact that I touched compilation dependencies in the JitInterface but didn't update the scanner should have been a red flag. Fixes a Release-only issue discovered in #103. Can't wait to have the CI fully up and running.