ARM64: Add more cfi data (#8336)
While the assembler helpers follow the ARM64 ABI and build a correct chain of frame information the debugger still have trouble to show a correct call stack when it include one of these assembler functions. The callstack either stops with with the last of these functions or shows it like it was called recursive It although makes stepping really slow. 

This fix adds additional cfi data for each frame that the assembler code adds on the stack. With this additional information the debugger can now show a correct callstack.