Add Internal.Console
This is the same Internal.Console that CoreCLR has, but extended with Unix support.

Is there any appetite to have this, potentially in the shared partition?

As I was debugging the timer/threadpool issue on Linux, I quickly found out that due to timing, printf debugging was the only feasible way to find out what's going on.