Fix shared library build and add test
`/wholearchive:` is the voodoo necessary for linker to include object files from a library that are otherwise not referenced. We need to include the bootstrapper because it has an initializer to initialize runtime that we depend on.

I'm keeping the other workaround (`IlcProcessEntrypoint`) in place because we need it due to the `main`/`WinMain` schism for `OutputType=Exe/WinExe`.