Add test for hardware intrinsics (#147)
The AOT precompilation of these has interesting aspects that were not tested before. This is a cleaned up and enhanced version of a test I used locally when developing the feature. Couldn't check it before in because the intrinsics didn't exist in the framework we compiled tests against in the CoreRT repo.