Fix behavior of interface dispatch on null instances (#239)
An interface call with a null `this` needs to throw a `NullReferenceException`. There are two ways to achieve that: either codegen can insert explicit code to trigger a NullRef at the callsite (before calling into the runtime library to do the dispatch), or the runtime library needs to be able to unwind the helper code that tries to resolve the dispatch and blame the callsite. .NET Native did the former, but the latter produces smaller code (no need to insert the checks). The CoreRT runtime doesn't have code to handle the latter. This pull request is adding that.

The fix is relatively straightforward - we mark locations in the helper where AV can occur. If an AV occurs outside managed code, we check whether it's one of the known locations. If so, we remove the helper method frame off the stack and pretend the AV happened in the calling code.

The one-liner fix in CorInfoImpl fixes an unintialized value problem that basically made us roll a dice on whether to do an explicit nullcheck at the callsite.

Fixes #221.