Add helper class to read trimming dataflow annotations
FlowAnnotations.cs corresponds to https://github.com/mono/linker/blob/master/src/linker/Linker.Dataflow/FlowAnnotations.cs.

This is a helper class that makes it possible to query for the new .NET 5 dataflow annotations within the compiler. We also implement warnings/validations when these annotations are incorrect.

This is mostly a copy-paste of FlowAnnotations from linker with tweaks to make it compile - the warning messages and error codes are identical. We want them to be identical to make suppressions work later.