Hook up dataflow analysis into compilation (#347)
With this, we can analyze reflection use using the dataflow analysis attributes introduced in .NET 5 (`DynamicallyAccessedMembersAttribute`).

We introduce a new node type (`DataflowAnalyzedMethodNode`) to make sure every method only gets analyzed once. The dependencies discovered by dataflow analysis are the depednencies of this node.

The node is injected into the graph by the metadata manager if we're running in the mode that collects metadata. Codegen backends (and scanners) need to call into this whenever they see a write to a field, or a method call. With RyuJIT, we don't have such granularity, so we also look at field loads and delegate creation sequences. The metadata manager only injects the dataflow analysis dependency if it sees the called method/written field are "interesting". This way we don't have to analyze many method bodies.