Cache result of MightHaveDispatchMap
We call this from multiple places within EEType construction, some of which are executed multiple times (especially the occurence within `ComputeOptionalEETypeFields` that is called way more often than it really should be). We should just call once and cache the result. Computing it is expensive because it requires resolving all interface methods on the type to their implementation.