Initial default interface method support
Support for default interface methods in the compiler and runtime.

It felt like the most straightforward approach is to just give interface types a dispatch map and populate vtables of interface method with the default implementation. I didn't realize they get a full vtable with a bunch of null slots. This might be something we could potentially optimize, but it's orthogonal.

There's plenty of things missing still, but it's a start and tests are starting to pass.