Add support for dataflow analysis to generic parameters (#372)
This will ensure that if e.g. a T in a generic type is annotated as "keep all methods on all the T's that this type is instantiated with", the compiler will do that.

What's still missing is validation that the generic annotations are consistent (e.g. if I have `Foo<[DynamicallyAccessed(PublicMethods] T>` and then I do `class Other<TOther> { Blah() => new Foo<TOther>()`, I will get a warning because TOther is not annotated the same as the T in Foo and we potentially broke the promise.

I think we can do this by hooking into the generic lookup tracking logic, but I don't want to add that now because I haven't figured out how to test that (we need to test a warning gets emitted).