CI: Run debug configuration instead of checked
The tests match the build configuration of the runtime. We don't have explicit `_r` and `_d` csproj for the smoke tests - this means that we always run with optimizations enabled.

We need coverage for the smoke tests with optimizations off - so we either add extra CSPROJ files a la JIT tests that have _r and _d, or just run a debug build in the CI.