Fix Checked multifile builds
The leaf (test) assemblies are correctly compiled with optimizations turned on in checked builds, but the framework wasn't. This resulted in fun-to-debug differences in generic dictionary shapes between the code compiled as part of the framework and dictionary lookups generated by app code. Optimizations can remove the need for some slots in certain cases.