Don't allow generic inlining if context differs
The `contextFromMethodBeingCompiled` optimization doesn't apply if we're changing the generic context from "method dictionary" to "the type".

Not sure if this can lead to other problems, but at minimum it causes a problem for the generic dictionary of the generic method computed by the scanner - we don't expect "foreign entries" in the method dictionary.