Disable BinaryFormatter by default (#480)
https://github.com/dotnet/designs/blob/main/accepted/2020/better-obsoletion/binaryformatter-obsoletion.md

For now, it can be turned back on by opting in explicitly, but per the above doc we reserve the right to just hardcode it as always disabled ("This hardcoded behavior extends to all other novel application types that might be introduced after the .NET 5 timeframe, including mobile apps and other restricted environment configurations.").

Gets rid of almost half of trimming warnings on a hello world app. Saves around 300 kB on a hello world.