Add a couple more things under FEATURE_UNIVERSAL_GENERICS
Saves around 20 kB on a Hello World. There's probably more opportunities, e.g. in the calling convention converter that handles various delegate thunks (that are only interesting in USG case since delegates are not special in normal canon mode). We could also entertain the option of getting rid of the calling convention converter altogether, but it might be needed to solve problems around dynamically loaded generic default interface methods that require an instantiating stub so I'm not sure about that.