Implement ILLinkTrim.xml-based resource filtering
Adds support for the `resource` element in ILLinkTrim files. This is the facility IL Linker (in cooperation with framework ILLinkTrim files) uses to remove framework resource strings.

This replaces the existing mechansm in ILCompiler that was exposed with `--removefeature:FrameworkStrings`.

Required a bit of build hackery so that nativeot libraries include the ILLinkTrim embedded file that instructs the compiler to remove the resource if the right feature switch is passed. I'm not too happy about that.