Optimize variant interface implementation generation (#514)
Saves 106 kB on a Hello World app.

Previously when the compiler saw we're generating a type that implements variant interfaces, we would throw hands in the air and generate the implementations of all variant interface methods on the type since we had no idea which ones would be actually needed.

This has drawbacks, especially around array types that implement a lot of variant (or variant-like) interfaces.

This commit introduces tracking of individual variant interface methods being used. It lets us avoid generating implementations of variant interface methods that are not statically seen as necessary.

We track this by adding conditional dependencies on a new `VariantInterfaceMethodUseNode` instead of the old logic that just generated hard `VirtualMethodUse` dependencies.