Check for basic block boundary before fetching arguments
Need to check for basic block boundary before we enter the for loop that looks for arguments of the conditional branch.

This handles the case when a conditional branch is a target of another branch. We might have incorrectly assumed that the flow of the conditional branch is a constant and replaced the unreachable branch of the conditional with a loop.