Fix problems with optimized builds of methods that `catch(T)`
We don't currently support this (tracked as https://github.com/dotnet/runtimelab/issues/204). The compiler compiles methods that do it into a `throw`.

But the way this is handled in CorInfoImpl is that we first compile the method and then invalidate it. The scanner on the other hand doesn't scan the method at all.

This results in a problem because during the compilation fo the method we're going to ask questions the scanner didn't compute.

It should be fine for scanner to just scan such methods. Alternatively, we could abort compilation earlier, but that's in code shared with crossgen2 and I don't want to touch it unnecessarily.