Annotate System.Private.Reflection.Core
I've analyzed all of nativeaot SDK libraries for trimming. Except for things that are already known in dotnet/runtime, we don't have warnings anymore after this.

(For future reference - to analyze trim safety of a library, one just need to compile the library in the `--multifile` mode and enable `--scanreflection`. This makes sure all code in the library is looked at.)