Limit impact of `typeof()` in a custom attribute (#620)
When a type is referenced by a typeof in a custom attribute, don't immediately create a constructed EEType for it.

This is a heuristic - we don't know what the type will be used for. The field/property/constructor parameter that is initialized with the `typeof` should ideally have a dataflow annotation that describes the intent - this is then going to root the appropriated structures.