Strip likely unreachable instance method bodies
Instance method bodies that can only be called with a null this (since we never allocated their owning type) are likely unreachable and can be stripped. C# would call them all through callvirt. Calling a method with a null this leads to general unhappiness anyway.

```csharp
object foo = ...;
if (foo is Foo f) // this condition is never actually true, but we don't eval that
    f.DoSomethingExpensive();
```

What this does:

* New node type that we give out instead of the real method body when someone is asking for a non-valuetype instance methods entrypoint.
* The new node works similar as ConstructedEEType/NecessaryEEType - it goes out of the way and acts as the real method body at object emission phase if the real body is part of the dependency graph.
* The fake body just tail calls into a throw helper so that's the experience someone would have if they actually call something with a null this.
* Also had to do something about inlining - since scanner also does this optimization, we had to prevent compilation phase from inlining the unreachable code. This is done by exporting this knowledge out of the scanner.