Add debug header for diagnostic tools (#709)
This change adds the concept of a "Debug Header" which allows diagnostic tooling to get information about the runtime types and layout without being tightly coupled to the runtime like the DAC is for coreclr. It is a lightweight self describing metadata format focused on types and their members, globals, and build defines.