More efficient debug loc info generation
This puts locinfo generation on the same plan as varinfo - only capture the native information in JitInterface and translate to managed terms at a later time.

Drops the number of total GC allocations by 10%. The throughput impact is mostly a wash since we no longer do the mapping in the multithreaded phase. This should be fixable, maybe even while we're still using the LLVM-based ObjWriter.