Implement better GetExceptionForHRInternal
This is taken from Mono.

We have an implementation in .NET Native that seems to be more complete but requires porting more code. We can probably get by with Mono's "good enough".

We could alternatively block tests that test this on COM support. There's a bunch of them. Not sure what's the point of this API without COM.