Implement Module.GetField
Long time ago in .NET Native, the module types were getting dropped on the floor during ILMerging. We don't do that here and the compiler generates all the metadata for it.

Only implementing fields because that's all we apparently test. Not signing up for writing new tests (filed an issue) or writing untested code.