Switch attribute equality to dotnet/runtime plan (#847)
.NET Native injected methods into these to make equality not reflection based. dotnet/runtime just walks fields using reflection.

It's about the equal amount of pain to deal with either one. The reflection based implementation should work for the mainstream framework scenarios. It will become more tricky if we want to make reflection metadata more pay for play in the future. The IL Linker suppressions on this implementation assume we need field metadata to operate. That's not the case in NativeAOT.