Add support for profile guided optimizations (#845)
Hooks up into PGO support in crossgen2 - the workflow is to generate profiling information with CoreCLR and consume it with the NativeAOT compiler.

Eventually, we should also add PGO data collection, but that's more than 2 hours of work.

I'm also adding the PGO data we use to build .NET 6 ReadyToRun assemblies to the package. This should be representative of some interesting workloads.