Update files shared with illink
This bring two major things:

* Ability to track contents of arrays in the dataflow analysis. We could later use this to model MakeGenericType and MakeGenericMethod.
* Support for DynamicallyAccessedMembersAttribute on classes and interfaces.

The semantic of DAMAttribute on classes and interface is: when a type has DAMAttribute, calls to `object.GetType` on locations statically typed as something having/inheriting the DAMAttribute return an annotated System.Type value.

The marking part of this required creating a new node in the dependency graph. The node represents "object.GetType" could have been called on this annotated type. It brings the dependencies mandated by the DAMAttribute.