Implement support for default interface methods
Different take from #368 - this time we're adding extra slots into the (sealed) vtable of the implementing class and pointing the dispatch map of the owning type to those slots when a default implementation is needed.

This strategy has a couple advantages:
* Main motivation: it will make implementing canonical interface methods possible - we will have a good spot to grab the interface type (instantiation argument of the default interface method) from, since we can now have per-implementing-type entrypoints.
* We are now able to figure out a default implementation is not needed at compile time (we don't generate a body unless there's a class that needs it)
* Dispatch is pre-computed (no longer have to hunt for the default implementation in dispatch maps of the interfaces at runtime).

The disadvantage is that this strategy is a bit more complicated, but it doesn't actually look too terrible.