Support for default interface methods on generic interfaces (#1122)
Add support for generating instantiating stubs. The implementation strategy is similar to how we do instantiating stubs on generic structs. Adds two kind of methods - the actual instantiating stub, and a fake method that is necessary so that we can access the instantiation argument from IL. The fake method is replaced with the real entrypoint when we ask for the reloc for it.