Add functionality to enumerate virtual methods only
We were spending around 8% of time in GetMethods due to various virtual method resolution code. In .NET, not that many methods are virtual, so restricting the iteration to virtual methods can save significant amount of time. The type system didn't have a facility to do this, so I'm adding one.

I'm seeing ~8% wallclock time savings when compiling a Hello World with this.