Duplicate value when loading from locations
We were trying to save memory allocations by keeping the Value used at the storage location (instead of allocating a fresh copy to put on the stack). This is problematic because the values on the stack are immutable, but the underlying locations are mutable. This is observable if the value is loaded on the evaluation stack and we then assign a new value to the location - the stack slot would change as well.

The allocation is inevitable.

Also fixes an off-by-one range check error.