Use dependent handle for implicit locks
The current implementation of implicit locks can keep the Lock and the associated event alive
for a long time after the owning object is dead, It can appear as a leak when observing
the performance metrics of the process.

Switch to managing the implicit locks lifetime using dependent handles. It has characteristics
that are much closer to CoreCLR implementation. The sync block index used by CoreCLR implementation
is effectively a dependent handle with extra hardcoded behavior. We may want to switch to the same
plan as CoreCLR eventually, but this change should improve the situation for now.