Make PropertyValue public (#1350)
ReferenceTypeHelper is a target of reflection. Likely none of that worked at runtime.

Noticed this by accident in the "generate template type metadata lazily" mode - I didn't subject the template type metadata to reflection blocking, and I was seeing extra template metadata for this compared to baseline. In the baseline, `ReferenceTypeHelper` is unusable with reflection completely - in the lazy mode, we generated unreachable template type metadata for it (and still blocked the reflection data).