NativeAOT-LLVM: Support for calling methods with return value in spill slot (#1339)
This PR copies some of the logic from IL->LLVM over to the RyuJit module for calling methods where the return value goes in a spill slot so that the conservative GC can track it. It does not add support for compiling methods which themselves have return values that need a spill slot.