NativeAOT-LLVM: Support for non void LLVM methods (#1358)
This PR starts to add support in the RyuJIT module for calling methods where the return value can be returned on the LLVM stack, i.e. through the ret instruction. Just ints handled in this PR. Also fixes a bug when the only use of the statics symbol was from non-GC fields, then then symbol was not expanded.

Also widen some LLVM int operations, e.g. ICmpNE which returns an i8 to i32 at the point of storing the result in a local.

* add gc statics symbol for GetNonGCStaticBase

* better variable name.

* map bools to i8, not i1