Delete CheckArrayElementType lookup support
This appears unused. It was introduced in 2014 for what I assume was lazy generic support in rhbind. ProjectX didn't need it for anything so likely won't be needed in NativeAOT either. This helper is always the same so I don't think we'll need to put it in a generic dictionary.