Change how we write to stderr from CoreLib
We write to stderr from CoreLib for two reasons:

1. FailFast
2. Unhandled exception (also FailFast in a way)

We're using a roundabout way to use `Console.Error.WriteLine` but that has problems for FailFast because we then don't fail fast enough (#1277).

This inlines enough of writing to stderr that we can get rid of System.Console dependency.

This is complicated on Windows because output can be redirected and then we need to deal with code pages. System.Console uses the encoding of the Console even when we're redirected to a file/pipe. I'm making this go through the same infrastructure. If output is not redirected, I'm just calling into `WriteConsole` directly, without having to mess with encodings.

Fixes #1277.