Include canonical EETypes in GVM analysis
We might no longer have all the concrete EETypes to do analysis - need to include the canonical EETypes we have in the system. The main part of the change is the move of `InterestingForDynamicDependencyAnalysis` from `ConstructedEETypeNode` to `EETypeNode` and limiting it so that it only applies to Constructed and Canonical EETypes (that's the new if at the beginning).

The rest is just mopping up after the change.

Fixes #1372