Implement IDynamicInterfaceCastable
Adds almost full support for IDynamicInterfaceCastable. The missing part is support for shared interface methods - we need to provide generic context somehow and instantiating stubs are not it because we don't have them.

I used https://github.com/dotnet/corert/pull/8103 as a blueprint, with a small difference - instead of hardcoding slot numbers into the MethodTable, I'm making runtime call into CoreLib so that we can call methods on `IDynamicInterfaceCastable` as usual. Being in the CoreLib makes various error handling easier too.

The implementation in the compiler is very straightforward:

* Allow `DynamicInterfaceCastableImplementation` interfaces to have a dispatch map.
* Allow `DynamicInterfaceCastableImplementation` to have a sealed vtable.
* Pretty much all that's needed.

At runtime, we consult the dispatch map to find the proper implementation, given an implementation interface.